feat(monitoring): update monitoring.

+++ Backend-farmlogitech/Monitoring/Domain/Repositories/Messages/IMessageRepository.cs
@@ -5,8 +5,10 @@ namespace Backend_farmlogitech.Monitoring.Domain.Repositories;
 
 public interface IMessageRepository: IBaseRepository<Message>
 { 
-    Task<IEnumerable<Message>> FindAllByCollaboratorId(long collaboratorId);
-    Task<IEnumerable<Message>> FindAllByCollaboratorIdAndFarmerId(long collaboratorId, long farmerId);
-    Task<IEnumerable<Message>> FindAllMessageByCollaboratorIdAndTransmitterIdNot(long collaboratorId, long transmitterId);
-    Task<IEnumerable<Message>> FindAllMessageByFarmerIdAndTransmitterIdNot(long farmerId, long transmitterId);
+    Task<IEnumerable<Message>> FindAllByCollaboratorId(int collaboratorId);
+    Task<IEnumerable<Message>> FindAllByCollaboratorIdAndFarmerId(int collaboratorId, int farmerId);
+    Task<IEnumerable<Message>> FindAllMessageByCollaboratorIdAndTransmitterIdNot(int collaboratorId, int transmitterId);
+    Task<IEnumerable<Message>> FindAllMessageByFarmerIdAndTransmitterIdNot(int farmerId, int transmitterId);
 }
+
+
